flink_job/requirements.txt
pymongo==4.6.1
kafka-python==2.0.2